AI_Webscraper/requirements.txt
selenium
requests
openai
webdriver_manager
beautifulsoup4
langchain
lxml
html5lib
pandas
pydantic
langchain_openai
langchain_community
html2text
readability-lxml
openpyxl
openai
python-dotenv
streamlit
streamlit-tags
groq      